python-server/requirements.txt
python-socketio
eventlet
flask-cors